requirements.txt
pandas
mysql-connector-python
boto3
apache-airflow
kagglehub

